Add I915 minigbm backend for chromeos support
Test: build script works and shows pixels on chromeos

+++ rebuild.sh
@@ -15,15 +15,15 @@ DEFINE_string arm_system \
   "" "IP address or DNS name of an ARM system to do the secondary build"
 DEFINE_string arm_user \
   "vsoc-01" "User to invoke on the ARM system"
-DEFINE_string build_instance \
-  "${USER}-build" "Instance name to create for the build" "i"
-DEFINE_string build_user cuttlefish_crosvm_builder \
-  "User name to use on GCE when doing the build"
 DEFINE_string project "$(gcloud config get-value project)" "Project to use" "p"
 DEFINE_string source_image_family debian-9 "Image familty to use as the base" \
   "s"
 DEFINE_string source_image_project debian-cloud \
   "Project holding the base image" "m"
+DEFINE_string x86_instance \
+  "${USER}-build" "Instance name to create for the build" "i"
+DEFINE_string x86_user cuttlefish_crosvm_builder \
+  "User name to use on GCE when doing the build"
 DEFINE_string zone "$(gcloud config get-value compute/zone)" "Zone to use" "z"
 
 SSH_FLAGS=(${INTERNAL_IP})
@@ -52,54 +52,56 @@ main() {
   if [[ "${fail}" -ne 0 ]]; then
     exit "${fail}"
   fi
-  project_zone_flags=(--project="${FLAGS_project}" --zone="${FLAGS_zone}")
-  delete_instances=("${FLAGS_build_instance}")
-  gcloud compute instances delete -q \
-    "${project_zone_flags[@]}" \
-    "${delete_instances[@]}" || \
-      echo Not running
-  gcloud compute instances create \
-    "${project_zone_flags[@]}" \
-    --boot-disk-size=200GB \
-    --machine-type=n1-standard-4 \
-    --image-family="${FLAGS_source_image_family}" \
-    --image-project="${FLAGS_source_image_project}" \
-    "${FLAGS_build_instance}"
-  wait_for_instance "${FLAGS_build_instance}"
-  # beta for the --internal-ip flag that may be passed via SSH_FLAGS
-  gcloud beta compute scp "${SSH_FLAGS[@]}" \
-    "${project_zone_flags[@]}" \
-    "${source_files[@]}" \
-    "${FLAGS_build_user}@${FLAGS_build_instance}:"
-  gcloud compute ssh "${SSH_FLAGS[@]}" \
-    "${project_zone_flags[@]}" \
-    "${FLAGS_build_user}@${FLAGS_build_instance}" -- \
-      ./rebuild_gce.sh
-  gcloud beta compute scp --recurse "${SSH_FLAGS[@]}" \
-    "${project_zone_flags[@]}" \
-    "${FLAGS_build_user}@${FLAGS_build_instance}":x86_64-linux-gnu \
-    "${ANDROID_BUILD_TOP}/device/google/cuttlefish_vmm"
-  gcloud beta compute scp "${SSH_FLAGS[@]}" \
-    "${project_zone_flags[@]}" \
-    "${FLAGS_build_user}@${FLAGS_build_instance}":clean-source.tgz \
-    "${ANDROID_BUILD_TOP}/device/google/cuttlefish_vmm"
-  gcloud compute disks describe \
-    "${project_zone_flags[@]}" "${FLAGS_build_instance}" | \
-      grep ^sourceImage: > "${DIR}"/x86_64-linux-gnu/builder_image.txt
+  if [[ -n "${FLAGS_x86_instance}" ]]; then
+    project_zone_flags=(--project="${FLAGS_project}" --zone="${FLAGS_zone}")
+    delete_instances=("${FLAGS_x86_instance}")
+    gcloud compute instances delete -q \
+      "${project_zone_flags[@]}" \
+      "${delete_instances[@]}" || \
+        echo Not running
+    gcloud compute instances create \
+      "${project_zone_flags[@]}" \
+      --boot-disk-size=200GB \
+      --machine-type=n1-standard-4 \
+      --image-family="${FLAGS_source_image_family}" \
+      --image-project="${FLAGS_source_image_project}" \
+      "${FLAGS_x86_instance}"
+    wait_for_instance "${FLAGS_x86_instance}"
+    # beta for the --internal-ip flag that may be passed via SSH_FLAGS
+    gcloud beta compute scp "${SSH_FLAGS[@]}" \
+      "${project_zone_flags[@]}" \
+      "${source_files[@]}" \
+      "${FLAGS_x86_user}@${FLAGS_x86_instance}:"
+    gcloud compute ssh "${SSH_FLAGS[@]}" \
+      "${project_zone_flags[@]}" \
+      "${FLAGS_x86_user}@${FLAGS_x86_instance}" -- \
+        ./rebuild_gce.sh
+    gcloud beta compute scp --recurse "${SSH_FLAGS[@]}" \
+      "${project_zone_flags[@]}" \
+      "${FLAGS_x86_user}@${FLAGS_x86_instance}":x86_64-linux-gnu \
+      "${ANDROID_BUILD_TOP}/device/google/cuttlefish_vmm"
+    gcloud beta compute scp "${SSH_FLAGS[@]}" \
+      "${project_zone_flags[@]}" \
+      "${FLAGS_x86_user}@${FLAGS_x86_instance}":clean-source.tgz \
+      "${ANDROID_BUILD_TOP}/device/google/cuttlefish_vmm"
+    gcloud compute disks describe \
+      "${project_zone_flags[@]}" "${FLAGS_x86_instance}" | \
+        grep ^sourceImage: > "${DIR}"/x86_64-linux-gnu/builder_image.txt
+  fi
   if [[ -n "${FLAGS_arm_system}" ]]; then
     scp \
       "${source_files[@]}" \
       "${ANDROID_BUILD_TOP}/device/google/cuttlefish_vmm/clean-source.tgz" \
       "${FLAGS_arm_user}@${FLAGS_arm_system}:"
-    ssh "${FLAGS_arm_user}@${FLAGS_arm_system}" -- \
+    ssh -t "${FLAGS_arm_user}@${FLAGS_arm_system}" -- \
         ./rebuild_gce.sh secondary_build
-    scp -r "${FLAGS_build_user}@${FLAGS_build_instance}":arm64-linux-gnu \
+    scp -r "${FLAGS_arm_user}@${FLAGS_arm_system}":aarch64-linux-gnu \
       "${ANDROID_BUILD_TOP}/device/google/cuttlefish_vmm"
   fi
   exit 0
   gcloud compute instances delete -q \
     "${project_zone_flags[@]}" \
-    "${FLAGS_build_instance}"
+    "${FLAGS_x86_instance}"
 }
 
 FLAGS "$@" || exit 1
